Split out SneakView
